backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-generativeai
google-cloud-storage
google-cloud-logging
google-cloud-secret-manager
google-cloud-aiplatform
pydantic
python-dotenv
cors
pytest
pytest-asyncio
